rename stuff to fit u8, u16, u32, s32 convention

+++ bluespider/gui_connections.py
@@ -73,7 +73,7 @@ def make_header_connections(self):
                 if data_element in structure:
                     size = structure[data_element][0]
                 else: # Bah, don't check it (it'll apply only to signposts)
-                    size = "long"
+                    size = "u32"
                 if not mapped.fits(num, size):
                     raise Exception(data_element + " too big")
                 if size == "ptr" and num < 0x8000000 and num != 0:
